requirements.txt
web3
python-dotenv
tqdm
requests
